refactor: split local journey workflow seams
